+++ src/modules/refresh.js
@@ -1,0 +1,10 @@
+import storage, { mainArr } from './storage.js';
+import addHtml from './addHtml.js';
+
+function refreshBtn() {
+  mainArr.splice(0, mainArr.length);
+  storage();
+  addHtml();
+}
+
+export default refreshBtn;
