+++ src/modules/storage.js
@@ -1,0 +1,7 @@
+const mainArr = JSON.parse(localStorage.getItem('array')) || [];
+
+function storage() {
+  localStorage.setItem('array', JSON.stringify(mainArr));
+}
+export default storage;
+export { mainArr };
